+++ R/Function_SinkFill.r
@@ -6,36 +6,40 @@ SinkFill <-
         if ( missing( DEM ) || !is( DEM, "RasterLayer" ) )
         { 
             stop("Invalid type of DEM") 
+        }else if ( xres( DEM ) == yres( DEM ) )
+        {
+            # Extract basic infomation of a DEM
+            res     <- xres( DEM )
+            xmn     <- extent(DEM)[1]
+            xmx     <- extent(DEM)[2]
+            ymn     <- extent(DEM)[3]
+            ymx     <- extent(DEM)[4]
+            crs     <- proj4string(DEM)
+            # Convert format of a DEM from raster to matrix for FORTRAN subroutine.  
+            DEM[is.na(DEM)] <- -999999 
+            DEM_m   <- as.matrix( DEM )
         }else{
-            if ( xres( DEM ) == yres( DEM ) )
-            {
-                DEM_m   <- as.matrix( DEM )
-                res     <- xres( DEM )
-            }else{
-                stop("Resolution of DEM is not compatible")
-            }
+            stop("Resolution of DEM is not compatible")
         }
 
         # 2. Boundary check
         if ( missing( Boundary ) )
         { 
-            Boundary_list <- .Fortran(C_checkboundary, DEM = DEM_m, nr = nrow(DEM), nc=ncol(DEM), boundary = DEM_m*0, NAOK=T)
+            Boundary_list <- .Fortran(C_checkboundary, DEM = DEM_m, nr = nrow(DEM), nc=ncol(DEM), boundary = DEM_m, NAOK = T)
             Boundary_m <- Boundary_list$boundary 
-        }else{
-            Boundary_m <- as.matrix( Boundary ) 
+            Boundary <- raster(Boundary_m, xmn = xmn, xmx = xmx, ymn = ymn, ymx = ymx, crs = crs)
         }
-        
-        # 3. Result map
-        DEM_nosink_m <- 0 * DEM_m
+        Boundary[is.na(Boundary)] <- -999999 
+        Boundary_m   <- as.matrix( Boundary )
 
         # Run SinkFill module
         SinkFill_result <- .Fortran( C_sinkfill, DEM = DEM_m, nr = nrow( DEM ), 
                                     nc = ncol( DEM ), res = res, boundary = Boundary_m,
-                                    min_angle = min_angle, DEM_nosink = DEM_nosink_m,
+                                    min_angle = min_angle, DEM_nosink = DEM_m,
                                     NAOK = T )
-        
+
         # Output
-        DEM_nosink <- DEM; DEM_nosink[] <- SinkFill_result$DEM_nosink
+        DEM_nosink <- raster(SinkFill_result$DEM_nosink, xmn = xmn, xmx = xmx, ymn = ymn, ymx = ymx, crs = crs)
         return( DEM_nosink )
     }